Proposal for using engines as objects rather than classes.
