+++ seahorse/lib/seahorse/json/engines/oj_engine.rb
@@ -9,16 +9,24 @@ module Seahorse
       # @api private
       class OjEngine
 
-        def self.load(json)
-          ::Oj.load(json, mode: :compat, symbol_keys: false, empty_string: false)
+        DEFAULT_LOAD_OPTS = { mode: :compat, symbol_keys: false, empty_string: false }.freeze
+
+        DEFAULT_DUMP_OPTS = { mode: :compat }.freeze
+
+        def initialize(load_opts: {}, dump_opts: {})
+          @load_opts = DEFAULT_LOAD_OPTS.merge(load_opts)
+          @dump_opts = DEFAULT_DUMP_OPTS.merge(dump_opts)
+        end
+
+        def load(json)
+          ::Oj.load(json, @load_opts)
         rescue ::Oj::ParseError, EncodingError, JSON::ParserError => e
           raise ParseError, e
         end
 
-        def self.dump(value)
-          ::Oj.dump(value, mode: :compat)
+        def dump(value)
+          ::Oj.dump(value, @dump_opts)
         end
-
       end
     end
   end
